Loki Logs sample: use the selected time range in sample request (https://github.com/grafana/grafana/pull/65126)
Logs sample: use the selected time range in sample request

There are parts of the query, such as label values, that could only be present during the current selected time interval. If we request using a time range other than the selected one, we are not providing builder suggestions based on the same data that the user is seeing.